情報をシェアする: Rename components/DataViewShare.vue to components/index/_shared/DataView/Share.vue
